+++ WorQit/WorQit/Vacancies.xaml.cs
@@ -4,6 +4,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Net.Http;
 using System.Threading.Tasks;
+using Windows.Devices.Geolocation;
+using Windows.Services.Maps;
 using Windows.UI.Popups;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
@@ -21,6 +23,7 @@ namespace WorQit
         private int currentVacancyIndex = 0;
         private Vacancy currentVacancy;
         private int progressValueUpdate;
+        Windows.Storage.ApplicationDataContainer localSettings = Windows.Storage.ApplicationData.Current.LocalSettings;
 
         public Vacancies()
         {
@@ -35,7 +38,7 @@ namespace WorQit
         {
             await getVacancies();
             getCurrentHighestVacancy();
-            if(matchedList.Count != 0)
+            if (matchedList.Count != 0)
             {
                 setProgress();
             }
@@ -43,15 +46,32 @@ namespace WorQit
 
         public async Task getVacancies()
         {
-            using (var client = new System.Net.Http.HttpClient())
+            using (var client = new HttpClient())
             {
-                var uri = new Uri("http://worqit.azurewebsites.net/api/Vacancy/getVacanciesByScore/" + Login.loggedInUser.ID.ToString());
+                string distance = "";
+                string hours = "";
+                string salary = "";
+                try
+                {
+                    distance = (string)localSettings.Values["distance"];
+                    hours = (string)localSettings.Values["hours"];
+                    salary = (string)localSettings.Values["salary"];
+                }
+                catch(Exception ex)
+                {
+
+                }
+                var uri = new Uri("http://worqit.azurewebsites.net/api/Vacancy/getVacanciesByScore/" + Login.loggedInUser.ID.ToString() + "?salary=" + salary + "&hours=" + hours);
                 var response = await client.GetAsync(uri);
                 var result = await response.Content.ReadAsStringAsync();
                 var vacanciesRoot = JsonConvert.DeserializeObject<VacancyRootObject>(result);
-                foreach(var vacancy in vacanciesRoot.Vacancys)
+                foreach (var vacancy in vacanciesRoot.Vacancys)
                 {
-                    matchedList.Add(vacancy);
+                    double distance2 = await getDistance(vacancy.location);
+                    if (distance2 <= Double.Parse(distance))
+                    {
+                        matchedList.Add(vacancy);
+                    }
                 }
             }
         }
@@ -92,7 +112,7 @@ namespace WorQit
 
         public async void setRating(int empID, int vacID, int rating)
         {
-            using (var client = new System.Net.Http.HttpClient())
+            using (var client = new HttpClient())
             {
                 try
                 {
@@ -121,7 +141,7 @@ namespace WorQit
             prgresVacancies.Value = prgresVacancies.Value + progressValueUpdate;
         }
 
-   
+
 
 
 
@@ -166,5 +186,74 @@ namespace WorQit
         {
             Frame.Navigate(typeof(Start));
         }
+
+        public async Task<double> getDistance(string locatie)
+        {
+
+            //The addresses to geocode.
+            string addressToGeocode = Login.loggedInUser.location;
+            string addressToGeocode2 = locatie;
+
+            //Query hint
+            BasicGeoposition queryHint = new BasicGeoposition();
+            queryHint.Latitude = 47.643;
+            queryHint.Longitude = -122.131;
+            Geopoint hintPoint = new Geopoint(queryHint);
+
+            MapLocationFinderResult result =
+                  await MapLocationFinder.FindLocationsAsync(
+                                    addressToGeocode,
+                                    hintPoint,
+                                    3);
+
+            MapLocationFinderResult result2 =
+                  await MapLocationFinder.FindLocationsAsync(
+                                    addressToGeocode2,
+                                    hintPoint,
+                                    3);
+
+            var sCoord = new Position() { Latitude = result.Locations[0].Point.Position.Latitude, Longitude = result.Locations[0].Point.Position.Longitude };
+            var eCoord = new Position() { Latitude = result2.Locations[0].Point.Position.Latitude, Longitude = result2.Locations[0].Point.Position.Longitude };
+
+            var haversine = new Haversine();
+            double distance = haversine.Distance(sCoord, eCoord, DistanceType.Kilometers);
+            return distance;
+        }
+    }
+}
+
+public enum DistanceType { Miles, Kilometers };
+/// <summary>
+/// Specifies a Latitude / Longitude point.
+/// </summary>
+public struct Position
+{
+    public double Latitude;
+    public double Longitude;
+}
+public class Haversine
+{
+    /// <summary>
+    /// Returns the distance in miles or kilometers of any two
+    /// latitude / longitude points.
+    /// </summary>
+    public double Distance(Position pos1, Position pos2, DistanceType type)
+    {
+        double R = (type == DistanceType.Miles) ? 3960 : 6371;
+        double dLat = this.toRadian(pos2.Latitude - pos1.Latitude);
+        double dLon = this.toRadian(pos2.Longitude - pos1.Longitude);
+        double a = Math.Sin(dLat / 2) * Math.Sin(dLat / 2) +
+            Math.Cos(this.toRadian(pos1.Latitude)) * Math.Cos(this.toRadian(pos2.Latitude)) *
+            Math.Sin(dLon / 2) * Math.Sin(dLon / 2);
+        double c = 2 * Math.Asin(Math.Min(1, Math.Sqrt(a)));
+        double d = R * c;
+        return d;
+    }
+    /// <summary>
+    /// Convert to Radians.
+    /// </summary>
+    private double toRadian(double val)
+    {
+        return (Math.PI / 180) * val;
     }
 }
